Refactor: Enhance and fix page navigation in MainScreen
This commit refactors the page navigation logic in `MainScreen.dart`.

The `_currentPage` ValueNotifier's listener that animated page changes in the `PageView` have been removed.

Instead, the `onTap` callback of the `BottomNavigationBar` now directly animates the `_pageController` to the selected page. The `onPageChanged` callback on `PageView` has also been removed as it's no longer necessary.

The listener on `_pageController` that updated `_currentPage` based on page scrolling remains.